requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.8
django-debug-toolbar==1.9.1
django-endless-pagination==2.0
django-filter==1.1.0
django-js-asset==0.1.1
django-mathfilters==0.4.0
django-modeltranslation==0.12.1
django-mptt==0.9.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.6
django-widget-tweaks==1.4.1
idna==2.6
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
six==1.11.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
urllib3==1.22
